Add files via upload
Dataset Augmentation Toolkit

This illustrates my toolkit designed to augment datasets by generating multiple unique outputs for each input entry. 
It achieves this by creating an array of outputs at a range of temperatures for each input, thereby diversifying the dataset and increasing the amount of data per entry. This approach enhances the richness and variability of the dataset, which can be valuable for training models that require diverse input data and for better generalization.